% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{set_cache_rep}
\alias{set_cache_rep}
\title{répertoire de cache persistant}
\usage{
set_cache_rep(cache_rep = find_cache_rep())
}
\arguments{
\item{cache_rep}{(character) le répertoire}
}
\value{
rien
}
\description{
répertoire de cache persistant
}
\concept{source_data}
