% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setroot.R
\name{set_sourcoise_root}
\alias{set_sourcoise_root}
\title{Force root}
\usage{
set_sourcoise_root(root = NULL, quiet = TRUE)
}
\arguments{
\item{root}{(default \code{NULL}, character) path of the root}

\item{quiet}{(default \code{TRUE} boolean) displays messages}
}
\value{
root set (character)
}
\description{
\code{sourcoise()} mechanism to find root of the project automatically can be bypassed.
This function is equivalent to setting the \code{sourcoise.root} option, except for storage of cache at the level of the file. To allow thus behaviour, root should be set to \code{NULL}.
}
\examples{
dir <- tempdir()
set_sourcoise_root(dir)

}
