% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcoise_clear.R
\name{sourcoise_clear}
\alias{sourcoise_clear}
\title{Vide le cache}
\usage{
sourcoise_clear(
  what = sourcoise_status(root = root, prune = FALSE),
  root = NULL
)
}
\arguments{
\item{what}{(--) a tibble such as the one obtained by \code{sourcoise_status()}, possibly filtered}

\item{root}{pour forcer le root (non recommandé)}
}
\value{
la liste des fichiers supprimés
}
\description{
Vide le cache
}
\seealso{
Other sourcoise: 
\code{\link{sourcoise}()},
\code{\link{sourcoise_refresh}()},
\code{\link{sourcoise_reset}()},
\code{\link{sourcoise_status}()}
}
\concept{sourcoise}
