% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tools.R
\name{sourcoise_priority}
\alias{sourcoise_priority}
\title{Set priority for sourcoise}
\usage{
sourcoise_priority(path, priority = 10, root = getOption("sourcoise.root"))
}
\arguments{
\item{path}{the file for which priority is set}

\item{priority}{the level of priority (10 is the default)}

\item{root}{the root -- use only if you know what it does}
}
\value{
list of json created
}
\description{
The desingated file is set to the priority. Low priority means that when refreshing those files are groing
to be executed first; the reason being that this file is used by other and that you prefer it is refreshed before refreshing other.
However, when refreshing, subcall to sourcoise are executed, unless priority is set to 0
}
\details{
So, if you want to refresh it first and then not (because it is long to refresh), then, ste it to 0

Default priority is 10
}
