% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcoise_refresh.R
\name{sourcoise_refresh}
\alias{sourcoise_refresh}
\title{Refresh cache by executing sources selected}
\usage{
sourcoise_refresh(
  what = NULL,
  force_exec = TRUE,
  unfreeze = TRUE,
  quiet = FALSE,
  init_fn = getOption("sourcoise.init_fn"),
  root = NULL,
  log = "INFO",
  .progress = TRUE
)
}
\arguments{
\item{what}{(tibble) a tibble as generated by \code{sourcoise_status()}, possibly filtered, (defaut to \code{source_status()} )}

\item{force_exec}{(boolean) if \code{TRUE} code is executed, no matter what is cached}

\item{unfreeze}{(boolean) (default \code{TRUE}) when possible, unfreeze and uncache .qmd files in a quarto project when data used by those .qmd has been refreshed}

\item{quiet}{no message if TRUE}

\item{init_fn}{(function) (default NULL) execute a function}

\item{root}{(default \code{NULL}) force root to be set, instead of letting the function finding the root, for advanced uses}

\item{log}{(string) (default "INFO") log levels as in \code{logger::log_threshold()} (c("OFF", "INFO", ...)), comes with a performance cost}

\item{.progress}{(boolean) (default \code{TRUE}) displays a progression bar based on previous execution timings}
}
\value{
a list of r scripts (characters) executed, with timing and success
}
\description{
Refresh cache by executing sources selected
}
\examples{
fs::file_copy(
  fs::path_package("sourcoise", "ipch", "prix_insee.r"),
  "/tmp/prix_insee.r",
  overwrite = TRUE)
# Force execution (root is set explicitly here, it is normally deduced from project)
data <- sourcoise("prix_insee.r", root = "/tmp/", force_exec = TRUE)
# we then refresh all caches
sourcoise_refresh(root = "/tmp")
}
\seealso{
Other sourcoise: 
\code{\link{sourcoise}()},
\code{\link{sourcoise_clear}()},
\code{\link{sourcoise_reset}()},
\code{\link{sourcoise_status}()}
}
\concept{sourcoise}
