% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcoise_refresh.R
\name{sourcoise_refresh}
\alias{sourcoise_refresh}
\title{Exécute les sources sélectionnés}
\usage{
sourcoise_refresh(
  what = NULL,
  cache_rep = NULL,
  force_exec = TRUE,
  hash = TRUE,
  unfreeze = TRUE,
  quiet = TRUE,
  init_qmd = TRUE,
  root = root
)
}
\arguments{
\item{what}{un tibble issu de source_data (tout par défaut)}

\item{cache_rep}{le répertoire de cache si il n'est pas évident}

\item{force_exec}{(boléen) Si \code{TRUE} alors le code est exécuté (FALSE par défaut)}

\item{hash}{(boléen) (\code{TRUE} par défaut) vérifie les hashs}

\item{unfreeze}{(boléen) (\code{TRUE} par défaut) essaye de unfreezé et uncaché les qmd dont les données ont été rafraichies}

\item{quiet}{reste silencieux}

\item{init_qmd}{(\code{TRUE} par défaut) exécute \code{ofce::init_qmd()}}

\item{root}{(\code{NULL} par défaut) essaye de trouver le root à partir du point d'exécution ou utilise celui fournit}
}
\value{
la liste des sources exécutés
}
\description{
Exécute les sources sélectionnés
}
\seealso{
Other sourcoise: 
\code{\link{sourcoise}()},
\code{\link{sourcoise_clear}()},
\code{\link{sourcoise_reset}()},
\code{\link{sourcoise_status}()}
}
\concept{sourcoise}
