% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcoise_clear.R
\name{sourcoise_reset}
\alias{sourcoise_reset}
\title{Efface les .sourcoise}
\usage{
sourcoise_reset(root = NULL)
}
\arguments{
\item{root}{pour forcer le root (non recommandé)}
}
\description{
Efface les .sourcoise
}
\seealso{
Other sourcoise: 
\code{\link{sourcoise}()},
\code{\link{sourcoise_clear}()},
\code{\link{sourcoise_refresh}()},
\code{\link{sourcoise_status}()}
}
\concept{sourcoise}
