% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcoise_clear.R
\name{sourcoise_reset}
\alias{sourcoise_reset}
\title{Reset sourcoise}
\usage{
sourcoise_reset(root = NULL)
}
\arguments{
\item{root}{to force root (expert use)}
}
\description{
Removes all \code{.sourcoise} folders found under the project root.
}
\examples{
  fs::file_copy(
    fs::path_package("sourcoise", "ipch", "prix_insee.r"),
    "/tmp/prix_insee.r",
    overwrite = TRUE)
  data <- sourcoise("prix_insee.r", root = "/tmp/", force_exec = TRUE)
  sourcoise_reset(root = "/tmp/")
}
\seealso{
Other sourcoise: 
\code{\link{sourcoise}()},
\code{\link{sourcoise_clear}()},
\code{\link{sourcoise_refresh}()},
\code{\link{sourcoise_status}()}
}
\concept{sourcoise}
