% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcoise_status.R
\name{sourcoise_status}
\alias{sourcoise_status}
\title{Etat du cache de sourcoise}
\usage{
sourcoise_status(
  quiet = TRUE,
  root = NULL,
  src_in = getOption("sourcoise.src_in") \%||\% "project",
  prune = TRUE
)
}
\arguments{
\item{quiet}{(boléen) TRUE par défaut, ne dit rien}

\item{root}{force le root -- à ne pas utiliser sauf expert}

\item{src_in}{est ce que les données sont avec les qmd ?}

\item{prune}{(boléen) limite la liste au dernier cache (TRUE par défaut)}
}
\value{
tibble des fichiers en cache avec les informations des scripts appelants
}
\description{
Donne des informations sur le cache de source_data sous la forme d'un tibble
}
\seealso{
Other sourcoise: 
\code{\link{sourcoise}()},
\code{\link{sourcoise_clear}()},
\code{\link{sourcoise_refresh}()},
\code{\link{sourcoise_reset}()}
}
\concept{sourcoise}
